reverted last commit and moved to common
