+++ common/errors.go
@@ -7,3 +7,6 @@ var ErrInvalidTimeout = errors.New("invalid timeout value")
 
 // ErrNilArwenChangeLocker signals that a nil arwen change locker has been provided
 var ErrNilArwenChangeLocker = errors.New("nil arwen change locker")
+
+// ErrNilExclusionTree signals that a nil exclusion tree was provided
+var ErrNilExclusionTree = errors.New("nil exclusion tree")
