+++ factory/mock/processComponentsStub.go
@@ -53,7 +53,6 @@ type ProcessComponentsMock struct {
 	ESDTDataStorageHandlerForAPIInternal vmcommon.ESDTNFTStorageHandler
 	AccountsParserInternal               genesis.AccountsParser
 	ReceiptsRepositoryInternal           factory.ReceiptsRepository
-	HardforkExclusionHandlerField        process.HardforkExclusionHandler
 }
 
 // Create -
@@ -259,11 +258,6 @@ func (pcm *ProcessComponentsMock) ESDTDataStorageHandlerForAPI() vmcommon.ESDTNF
 // ReceiptsRepository -
 func (pcm *ProcessComponentsMock) ReceiptsRepository() factory.ReceiptsRepository {
 	return pcm.ReceiptsRepositoryInternal
-}
-
-// HardforkExclusionHandler -
-func (pcm *ProcessComponentsMock) HardforkExclusionHandler() process.HardforkExclusionHandler {
-	return pcm.HardforkExclusionHandlerField
 }
 
 // IsInterfaceNil -
